# improve performent of hash join in some case