@imay 2020-03-19
Hi, @yangzhg 
Could you attach some benchmark for this PR?